% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Alpha}
\alias{Alpha}
\title{Constructs alpha prior}
\usage{
Alpha(data, alpha, source, time = NULL, location = NULL)
}
\arguments{
\item{data}{long-format data.frame containing Dirichlet prior hyperparameter, source, time, and location columns}

\item{alpha}{name of hyperparameter column}

\item{source}{name of source column}

\item{time}{name of optional time column}

\item{location}{name of optional location column}
}
\value{
An Alpha_ data structure for use in sourceR models
}
\description{
The Alpha constructure function returns an R6 Alpha_
class which feeds sanitised prior or initialisation values for alpha into the model.
}
