% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alpha_}
\alias{Alpha_}
\title{Alpha prior hyperparameter class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Alpha_
}
\description{
Alpha prior hyperparameter class
}
\keyword{datasets}
