% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{DataNode}
\alias{DataNode}
\title{DataNode}
\format{
Object of \code{\link{R6Class}} with methods for constructing a DAG.
}
\value{
Object of \code{\link{DataNode}}
}
\description{
Represents a static data node in a DAG.
}
\details{
Derived from \link{Node}, please see base class documentation.
}
\section{Fields}{

\describe{
\item{\code{data}}{the data}
}}

\section{Methods}{

\describe{
  \item{\code{getData()}}{returns the node's data.}
  }
}

\keyword{DAG}
\keyword{node}
