% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{DirichletNode}
\alias{DirichletNode}
\title{DirichletNode}
\format{
Object of \code{\link{R6Class}} with methods for constructing a DAG.
}
\value{
Object of \code{\link{DirichletNode}}
}
\description{
Represents a d-dimensional Dirichlet distribution node in a DAG.
}
\details{
Derived from \link{StochasticNode}, please see base class documentation.
}
\section{Methods}{

\describe{
  \item{\code{new(data, alpha)}}{Create a DirichletNode with data vector \code{data} (length > 1) and parameter vector
   \code{alpha}}.
  }
}

\keyword{DAG}
\keyword{node}
