% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{DirichletProcessNode}
\alias{DirichletProcessNode}
\title{Transformed Dirichlet node}
\format{Object of \code{\link{R6Class}} with methods for constructing a DAG.}
\usage{
DirichletProcessNode
}
\value{
Object of \code{\link{DirichletProcessNode}}
}
\description{
Uses transformation due to Betancourt 2013
https://arxiv.org/pdf/1010.3436v4.pdf
}
\details{
Experimental -- do not use.
DirichletProcessNode

Represents a Dirichlet process as a single node in a DAG.

Derived from \link{StochasticNode}, please see base class documentation.
}
\section{Methods}{

\describe{
  \item{\code{new(theta, s, alpha, base, ...)}}{Create a DirichletProcessNode with value vector
  \code{theta} (length > 1), initial grouping vector \code{s}, concentration parameter \code{alpha}, and
  base distribution \code{base}.  Base should be a distribution function (dnorm, dgamma, etc) whose parameters
  are specified in \code{...}.}
  }
}
\keyword{DAG}
\keyword{node}

