% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{DPModel_impl}
\alias{DPModel_impl}
\title{Builds the source attribution model. Is not intended to be used by a regular user.
Developers only here!}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
Builds the source attribution model. Is not intended to be used by a regular user.
Developers only here!
}
\section{Fields}{

\describe{
\item{\code{y}}{3D array of [type, time, location] of the number of human cases}

\item{\code{X}}{3D array of the number of positive samples for each type, source and time
[type, source, time]}

\item{\code{R}}{3D array of normalised relative prevalences for each timepoint
[type, source, time]}

\item{\code{Time}}{a character vector of timepoint ids matching time dimension in y and R}

\item{\code{Location}}{a character vector of location ids matching location dimension in y}

\item{\code{Sources}}{a character vector of source ids matching the source dimension in X}

\item{\code{Type}}{a character vector of type ids matching the type dimension in X}

\item{\code{prev}}{a 2D array (matrix) of [source, time].}

\item{\code{a_q}}{concentration parameter for the DP}

\item{\code{a_theta}}{shape parameter for the Gamma base distribution for the DP}

\item{\code{b_theta}}{rate parameter for the Gamma base distribution for the DP}

\item{\code{a_r}}{3D array of [type, src, time] for the hyperprior on the relative prevalences R}

\item{\code{a_alpha}}{3D array of [source, time, location] for the prior on the alpha parameters}

\item{\code{s}}{vector giving initial group allocation for each type for the DP}

\item{\code{theta}}{vector giving initial values for each group in the DP}

\item{\code{alpha}}{3D array of [source, time, location] giving initial values for the alpha
parameters}
}}

