% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{FormulaNode}
\alias{FormulaNode}
\title{FormulaNode}
\format{
Object of \code{\link{R6Class}} with methods for constructing a DAG.
}
\value{
Object of \code{\link{FormulaNode}}
}
\description{
Represents a formula node in a DAG.  Inherit from this node to specify some kind of
formula within the DAG, e.g. a linear predictor and/or link function.  Override the
FormulaNode$getData() method to apply your own function.
}
\details{
Derived from \link{Node}, please see base class documentation.
}
\section{Methods}{

\describe{
  \item{\code{getData()}}{returns the node's transformed data.}
  }
}

\keyword{DAG}
\keyword{node}
