% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{GammaNode}
\alias{GammaNode}
\title{GammaNode}
\format{
Object of \code{\link{R6Class}} with methods for constructing a DAG.
}
\value{
Object of \code{\link{GammaNode}}
}
\description{
Represents a Gamma distribution node in a DAG.  Requires parent nodes for shape and rate respectively as
specified in \link{dgamma}.
}
\details{
Derived from \link{StochasticNode}, please see base class documentation.
}
\section{Methods}{

\describe{
  \item{\code{new(data, shape=1, rate=1)}}{Create a Gamma node with data \code{data} and Nodes
   \code{shape} and \code{rate} as specified in \link{dgamma}.}
  }
}

\keyword{DAG}
\keyword{node}
