% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{PoissonNode}
\alias{PoissonNode}
\title{PoissonNode}
\format{
Object of \code{\link{R6Class}} with methods for constructing a DAG.
}
\value{
Object of \code{\link{PoissonNode}}
}
\description{
Represents a Poisson distribution node in a DAG
}
\details{
Derived from \link{StochasticNode}, please see base class documentation.
}
\section{Methods}{

\describe{
  \item{\code{new(data, lambda, offset)}}{create a PoissonNode, with mean \link{Node} \code{lambda}, and offset \link{Node} \code{offset}.}
  }
}

\keyword{DAG}
\keyword{node}
