% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Prev}
\alias{Prev}
\title{Constructs prevalence data}
\usage{
Prev(data, prev, source, time = NULL)
}
\arguments{
\item{data}{long-format data.frame containing prevalence data by
source and time.}

\item{prev}{character string giving name of prevalence column in data}

\item{source}{character string giving name of source column in data}

\item{time}{optional column denoting times of prevalence observation}
}
\value{
A Prev data structure for use in sourceR models
}
\description{
The Prev constructor function returns an R6 Prevalence class
which feeds data into sourceR models.
}
