% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Q}
\alias{Q}
\title{Constructs initial values for q}
\usage{
Q(data, q, type)
}
\arguments{
\item{data}{long-format data.frame containing type-name and q-value for each observation.}

\item{q}{name of 'q' column}

\item{type}{name of type column}
}
\value{
A Q_ data structure for use in sourceR models
}
\description{
The Q constructor returns a R6 Q_ class which feeds sanitised
initial values for q into the model.
}
