% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SA_data.R
\docType{data}
\name{sim_SA_data}
\alias{sim_SA_data}
\title{Simulated data: Human cases of campylobacteriosis and numbers of source samples positive for \emph{Campylobacter}.}
\format{A data frame with 364 rows and 10 variables:
\describe{
  \item{Human}{number of human cases of campylobacteriosis}
  \item{Source1}{number of Source1 samples}
  \item{Source2}{number of Source2 samples}
  \item{Source3}{number of Source3 samples}
  \item{Source4}{number of Source4 samples}
  \item{Source5}{number of Source5 samples}
  \item{Source6}{number of Source6 samples}
  \item{Time}{Time id for the samples}
  \item{Location}{Location id for the samples}
  \item{Type}{MLST type id for the samples}
}}
\usage{
sim_SA_data
}
\description{
A simulated dataset containing the number of human cases of campylobacteriosis and numbers of source samples
positive for \emph{Campylobacter} for each bacterial subtype.
}
\keyword{datasets}

