% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SA_prev.R
\docType{data}
\name{sim_SA_prev}
\alias{sim_SA_prev}
\title{Simulated data prevalences.}
\format{A data frame with 12 rows and 3 variables:
\describe{
  \item{Value}{prevalence values}
  \item{Time}{Time id for the samples}
  \item{Source}{Source id for the samples}
}}
\usage{
sim_SA_prev
}
\description{
Source prevalences for the simulated data (sim_SA_data).
}
\keyword{datasets}

