% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sliceTensor}
\alias{sliceTensor}
\title{Slices a tensorA::tensor}
\usage{
sliceTensor(x, ...)
}
\arguments{
\item{x}{tensor array to be sliced}

\item{...}{arguments used to subset the sensor array}
}
\description{
Slices a tensorA::tensor, preserving the dimnames.
This is a workaround for a buggy implementation of
[.tensor as of tensorA v0.36.
}
