% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_ram.R
\name{source_ram}
\alias{source_ram}
\title{Allocate system RAM}
\usage{
source_ram(
  only_free = T,
  percent_of_total = 20,
  mem_unit = "GiB",
  round_up = T
)
}
\arguments{
\item{only_free}{If TRUE, only allocate the system's currently free memory.}

\item{percent_of_total}{An integer of what percent of memory to allocate, which also serves as an upper limit to allocating free ram.}

\item{mem_unit}{Currently assumes available memory is measured in GB -- More options TBD in a future version.}

\item{round_up}{If TRUE, it rounds decimals .5 and above up to the next whole number of GBs, otherwise rounds down.}
}
\value{

}
\description{
Allocate a portion of system RAM for R and/or Java to use.
}
