% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_wd.R
\name{source_wd}
\alias{source_wd}
\title{Configure the working directory}
\usage{
source_wd(this_file)
}
\arguments{
\item{this_file}{The name of the script calling the function. This uses here::i_am to set the project root correctly.}
}
\value{

}
\description{
Ensure the project root is set up correctly and create a global variable wd to represent the project root as working directory.
}
