\name{sourceR-package}
\alias{sourceR-package}
\alias{sourceR}
\docType{package}
\title{R code analysis tools}
\description{
  Static code analysis for R.
}

\details{
  cloneR consist of functions to visit R expressions objects to extract
  informations such as duplicated functions or naming convention.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\keyword{clone}
\keyword{code}
\keyword{analysis}
\keyword{detection}
