% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceR.R
\docType{package}
\name{sourceR}
\alias{sourceR}
\title{sourceR: A package for fitting Bayesian non-parametric source attribution models.}
\description{
The sourceR package currently provides an S6 class called HaldDP to fit the HaldDP model.
Further source attribution models are planned to be added to the package.
}
\section{sourceR methods}{

The main sourceR method is \code{\link{HaldDP}}.
}

