% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{StochasticNode}
\alias{StochasticNode}
\title{StochasticNode}
\format{
Object of \code{\link{R6Class}} with methods for constructing a DAG.
}
\value{
Object of \code{\link{StochasticNode}}
}
\description{
Represents a stochastic node in a DAG
}
\details{
Derived from \link{Node}, please see base class documentation.
}
\section{Fields}{

\describe{
\item{\code{data}}{contains the node's data}
}}

\section{Methods}{

\describe{
  \item{\code{logPosterior()}}{return the value of the log posterior distribution of the node.}
  \item{\code{getData()}}{returns the node's data.}
  }
}

\keyword{DAG}
\keyword{node}
