% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Y}
\alias{Y}
\title{Constructs disease count data}
\usage{
Y(data, y, type, time = NULL, location = NULL)
}
\arguments{
\item{data}{long-format data.frame containing source data}

\item{y}{character string giving name of disease counts column in data}

\item{type}{character string giving name of type column in data}

\item{time}{optional column denoting times of disease count observations}

\item{location}{optional column denoting location of disease count observations}
}
\value{
A Y disease count data structure for use in sourceR models
}
\description{
The Y constructor function returns an R6 Y class
which feeds disease count data into sourceR models.
}
