\name{sourceR-package}
\alias{sourceR-package}
\alias{sourceR}
\docType{package}
\title{
Fits a non-parametric source attribution model
}
\description{
This function implements a non-parametric source attribution model to attribute cases of disease to sources in Bayesian framework with source and type effects. Type effects are clustered using a Dirichlet Process. Multiple times and locations are supported. 
}

\details{
\tabular{ll}{
Package: \tab sourceR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-01\cr
License: \tab GPL-3\cr
}

}
\author{Poppy Miller <poppy.p.miller@gmail.com>}

%\Maintainer{Poppy Miller <poppy.p.miller@gmail.com>}

%\references{}
\keyword{ sourceR }
\keyword{ Bayesian }
\keyword{ non-parametric }
\keyword{ source attribution }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{}
