% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceR.R
\name{sourceR}
\alias{sourceR}
\title{Automagically source scripts}
\usage{
sourceR(
  source_dir = "R",
  include_rmd = F,
  exclude_dirs = "",
  exclude_files = "",
  recursive = T,
  exclude_marked_exclude = T,
  exclude_sourceR = T,
  echo_files_sourced = T
)
}
\arguments{
\item{source_dir}{The relative name of the directory within your project
folder containing scripts to source.}

\item{include_rmd}{TRUE to also source .Rmd files.}

\item{exclude_dirs}{A character string or vector of directories to exclude
from automatic sourcing.}

\item{exclude_files}{A character string or vector of files that should not be
sourced.}

\item{recursive}{TRUE to source files in the source_dir and any directories
within it, and so on.}

\item{exclude_marked_exclude}{TRUE to avoid sourcing any script with a
filename with an "_exclude" suffex, eg. "IncompleteScript_exclude.R"}

\item{exclude_sourceR}{TRUE to skip reloading this function from source
whenever the function is called.}

\item{echo_files_sourced}{TRUE to echo the list of files sourced to
output/console.}
}
\value{
Uses the source function to load and/or run scripts in the projects /R directory.
}
\description{
sourceR helps to manage project code by sourcing scripts selectively,
recursively. By default it will load any .R file in the project's /R
directory or subdirectories.
}
